improve dto, add controller tests

+++ packages/@n8n/api-types/src/dto/data-store/delete-data-store-rows.dto.ts
@@ -2,5 +2,5 @@ import { z } from 'zod';
 import { Z } from 'zod-class';
 
 export class DeleteDataStoreRowsDto extends Z.class({
-	ids: z.array(z.number().positive()),
+	ids: z.array(z.number().int().positive()),
 }) {}
